Remove redundant playlist component
