--- frontend/lyric-sync-app/src/components/YouTubePlaylistLoader/YouTubePlaylistLoader.scss
@@ -1,47 +1,0 @@
-@use '../../variables' as *;
-@use '../../mixins' as *;
-
-.youtube-playlist-viewer {
-  margin-top: 54px;
-  .load-section{
-    display: flex;
-    flex-direction: row;
-    .loading-spinner {
-      margin-left: 5px;
-      margin-top: 16px;
-      height: auto;
-    }
-  }
-
-  .error-message {
-    margin-top: 16px;
-    color: $error-color;
-  }
-
-  .data-grid-container {
-    margin-top: 16px;
-    height: 400px;
-    width: 100%;
-    border-radius: 8px;
-
-    .MuiDataGrid-row {
-      cursor: pointer;
-    }
-  }
-
-  .toolbar-title {
-    flex-grow: 1;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding-left: 16px;
-    padding: 24px;
-    border: 2px solid black;
-    border-radius: 8px;
-  }
-
-  .button-style {
-    background: $secondary-color;
-    border-radius: 8px;
-  }
-}
